requirements.txt
pandas==2.2.0
faker==23.0.0
kafka-python==2.0.2
dbt-postgres
python-logging-loki
